removed deprecated functions and renamed fileHandling -> persistence for better readability
