+++ src/diffcalc_API/services/Constraints.py
@@ -4,7 +4,7 @@ from diffcalc.hkl.constraints import Constraints
 
 from diffcalc_API.config import constraintsWithNoValue
 from diffcalc_API.errors.Constraints import check_constraint_exists
-from diffcalc_API.fileHandling import HklCalcRepo
+from diffcalc_API.persistence import HklCalcRepo
 
 
 async def get_constraints(name: str, repo: HklCalcRepo) -> str:
