+++ src/diffcalc_API/services/HklCalculation.py
@@ -9,7 +9,7 @@ from diffcalc_API.errors.HklCalculation import (
     check_valid_miller_indices,
     check_valid_scan_bounds,
 )
-from diffcalc_API.fileHandling import HklCalcRepo
+from diffcalc_API.persistence import HklCalcRepo
 
 PositionType = Tuple[float, float, float]
